# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Autopilot account, mode, and administrator API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface AuthSession { [key: string]: unknown }

export type ModeScenarioData = { [key: string]: unknown };

export interface Mode {
  id: string;
  adminId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @nullable */
  description?: string | null;
  scenarioData: ModeScenarioData;
  isActive?: boolean;
  isGlobal?: boolean;
  /** @nullable */
  targetUsers?: string[] | null;
  version?: number;
}

export type ScenarioInputScenarioData = { [key: string]: unknown };

export interface ScenarioInput {
  adminId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @nullable */
  description?: string | null;
  scenarioData: ScenarioInputScenarioData;
  isGlobal?: boolean;
}

export type ScenarioUpdateScenarioData = { [key: string]: unknown };

export interface ScenarioUpdate {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name?: string;
  /** @nullable */
  description?: string | null;
  scenarioData?: ScenarioUpdateScenarioData;
  isActive?: boolean;
}

export interface Profile {
  id: string;
  email: string;
  role: string;
  subscriptionStatus: string;
  /** @nullable */
  subscriptionExpiresAt?: string | null;
  isAdFree: boolean;
  adsWatchedToday: number;
  isBanned: boolean;
  /** @nullable */
  banReason?: string | null;
}

export interface AppSettings {
  id: number;
  /**
     * @minimum 1
     * @maximum 60
     */
  interstitialIntervalMinutes: number;
  /**
     * @minimum 1
     * @maximum 100
     */
  rewardAdsForOneDay: number;
  /**
     * @minimum 0
     * @maximum 30
     */
  trialDays: number;
}

export interface AppSettingsInput {
  /**
     * @minimum 1
     * @maximum 60
     */
  interstitialIntervalMinutes: number;
  /**
     * @minimum 1
     * @maximum 100
     */
  rewardAdsForOneDay: number;
  /**
     * @minimum 0
     * @maximum 30
     */
  trialDays: number;
}

export interface Subscription { [key: string]: unknown }

export type GrantSubscriptionBody = {
  /**
     * @minimum 1
     * @maximum 3650
     */
  durationDays: number;
  note?: string;
};

export type SetBannedBody = {
  banned?: boolean;
  reason?: string;
};

export type SetAdFreeBody = {
  adFree: boolean;
};