# lib/api-zod/src/generated/types/appSettingsInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Autopilot account, mode, and administrator API
 * OpenAPI spec version: 0.2.0
 */

export interface AppSettingsInput {
  /**
     * @minimum 1
     * @maximum 60
     */
  interstitialIntervalMinutes: number;
  /**
     * @minimum 1
     * @maximum 100
     */
  rewardAdsForOneDay: number;
  /**
     * @minimum 0
     * @maximum 30
     */
  trialDays: number;
}