# lib/api-zod/src/generated/types/authCredentials.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Autopilot account, mode, and administrator API
 * OpenAPI spec version: 0.2.0
 */

export interface AuthCredentials {
  email: string;
  /** @minLength 8 */
  password: string;
}