# lib/api-zod/src/generated/types/mode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Autopilot account, mode, and administrator API
 * OpenAPI spec version: 0.2.0
 */
import type { ModeScenarioData } from './modeScenarioData';

export interface Mode {
  id: string;
  adminId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @nullable */
  description?: string | null;
  scenarioData: ModeScenarioData;
  isActive?: boolean;
  isGlobal?: boolean;
  /** @nullable */
  targetUsers?: string[] | null;
  version?: number;
}