# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Autopilot account, mode, and administrator API
 * OpenAPI spec version: 0.2.0
 */

export interface Profile {
  id: string;
  email: string;
  role: string;
  subscriptionStatus: string;
  /** @nullable */
  subscriptionExpiresAt?: Date | null;
  isAdFree: boolean;
  adsWatchedToday: number;
  isBanned: boolean;
  /** @nullable */
  banReason?: string | null;
}