# lib/api-zod/src/generated/types/scenarioUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Autopilot account, mode, and administrator API
 * OpenAPI spec version: 0.2.0
 */
import type { ScenarioUpdateScenarioData } from './scenarioUpdateScenarioData';

export interface ScenarioUpdate {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name?: string;
  /** @nullable */
  description?: string | null;
  scenarioData?: ScenarioUpdateScenarioData;
  isActive?: boolean;
}